[MOVE/#13] 패키징 재설정
