+++ app/src/main/java/org/sopt/dosopttemplate/data/model/request/SignUpRequestDto.kt
@@ -1,14 +1,14 @@
-package org.sopt.dosopttemplate.data.datasource.model.response
+package org.sopt.dosopttemplate.data.model.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class LoginResponseDto(
-    @SerialName("id")
-    val id: Int,
+data class SignUpRequestDto(
     @SerialName("username")
     val username: String,
+    @SerialName("password")
+    val password: String,
     @SerialName("nickname")
     val nickname: String
 )
