+++ app/src/main/java/org/sopt/dosopttemplate/data/model/response/LoginResponseDto.kt
@@ -1,14 +1,14 @@
-package org.sopt.dosopttemplate.data.datasource.model.request
+package org.sopt.dosopttemplate.data.model.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class SignUpRequestDto(
+data class LoginResponseDto(
+    @SerialName("id")
+    val id: Int,
     @SerialName("username")
     val username: String,
-    @SerialName("password")
-    val password: String,
     @SerialName("nickname")
     val nickname: String
 )
