+++ app/src/main/java/org/sopt/dosopttemplate/data/service/FollowerService.kt
@@ -1,6 +1,6 @@
 package org.sopt.dosopttemplate.data.datasource.service
 
-import org.sopt.dosopttemplate.data.datasource.model.response.FollowerResponseDto
+import org.sopt.dosopttemplate.data.model.response.FollowerResponseDto
 import retrofit2.Call
 import retrofit2.http.GET
 import retrofit2.http.Query
