+++ app/src/main/java/org/sopt/dosopttemplate/di/ServicePool.kt
@@ -1,6 +1,6 @@
 package org.sopt.dosopttemplate.di
 
-import org.sopt.dosopttemplate.data.datasource.service.AuthService
+import org.sopt.dosopttemplate.data.service.AuthService
 import org.sopt.dosopttemplate.data.datasource.service.FollowerService
 
 object ServicePool {
