+++ app/src/main/java/org/sopt/dosopttemplate/presentation/auth/SignUpActivity.kt
@@ -7,12 +7,11 @@ import androidx.lifecycle.lifecycleScope
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import org.sopt.dosopttemplate.R
-import org.sopt.dosopttemplate.data.model.User
+import org.sopt.dosopttemplate.domain.entity.User
 import org.sopt.dosopttemplate.databinding.ActivitySignUpBinding
 import org.sopt.dosopttemplate.presentation.auth.LoginActivity.Companion.EXTRA_USER
 import org.sopt.dosopttemplate.util.base.BindingActivity
 import org.sopt.dosopttemplate.util.setOnSingleClickListener
-import org.sopt.dosopttemplate.util.toast
 import snackBar
 import toast
